Move utility parsers to utility namespace
