+++ src/parsers/domain/atomic_formula_skeleton.rs
@@ -1,7 +1,7 @@
 //! Provides parsers for atomic formula skeletons.
 
 use crate::parsers::domain::{parse_predicate, parse_variable};
-use crate::parsers::{parens, typed_list, ws};
+use crate::parsers::utility::{parens, typed_list, ws};
 use crate::types::domain::AtomicFormulaSkeleton;
 use nom::combinator::map;
 use nom::sequence::tuple;
