+++ src/parsers/domain/constants_def.rs
@@ -1,7 +1,7 @@
 //! Provides parsers for constant definitions.
 
 use crate::parsers::domain::parse_name;
-use crate::parsers::{prefix_expr, typed_list};
+use crate::parsers::utility::{prefix_expr, typed_list};
 use crate::types::domain::Constants;
 use nom::combinator::map;
 use nom::IResult;
