+++ src/parsers/domain/derived_predicate.rs
@@ -1,7 +1,7 @@
 //! Provides parsers for derived predicates.
 
 use crate::parsers::domain::{parse_atomic_formula_skeleton, parse_gd};
-use crate::parsers::prefix_expr;
+use crate::parsers::utility::prefix_expr;
 use crate::types::domain::DerivedPredicate;
 use nom::character::complete::multispace1;
 use nom::combinator::map;
