+++ src/parsers/domain/duration_constraint.rs
@@ -1,7 +1,7 @@
 //! Provides parsers for duration constraints.
 
 use crate::parsers::domain::parse_simple_duration_constraint;
-use crate::parsers::{prefix_expr, space_separated_list1};
+use crate::parsers::utility::{prefix_expr, space_separated_list1};
 use crate::types::domain::DurationConstraint;
 use nom::branch::alt;
 use nom::bytes::complete::tag;
