+++ src/parsers/domain/f_exp_da.rs
@@ -1,7 +1,7 @@
 //! Provides parsers for f-exp-da values..
 
 use crate::parsers::domain::{parse_binary_op, parse_f_exp, parse_multi_op};
-use crate::parsers::{parens, space_separated_list1};
+use crate::parsers::utility::{parens, space_separated_list1};
 use crate::types::domain::FExpDa;
 use nom::branch::alt;
 use nom::bytes::complete::tag;
