+++ src/parsers/domain/p_effect.rs
@@ -3,7 +3,7 @@
 use crate::parsers::domain::{
     atomic_formula, parse_assign_op, parse_f_exp, parse_f_head, parse_function_term, parse_term,
 };
-use crate::parsers::{parens, prefix_expr};
+use crate::parsers::utility::{parens, prefix_expr};
 use crate::types::domain::PEffect;
 use nom::branch::alt;
 use nom::bytes::complete::tag;
