+++ src/parsers/domain/timed_effect.rs
@@ -4,7 +4,7 @@ use crate::parsers::domain::{
     parse_assign_op_t, parse_cond_effect, parse_f_assign_da, parse_f_exp_t, parse_f_head,
     parse_time_specifier,
 };
-use crate::parsers::{parens, prefix_expr};
+use crate::parsers::utility::{parens, prefix_expr};
 use crate::types::domain::TimedEffect;
 use nom::branch::alt;
 use nom::character::complete::multispace1;
