+++ src/parsers/mod.rs
@@ -1,14 +1,4 @@
 //! Parser logic.
 
 pub mod domain;
-mod empty_or;
-mod typed_list;
-mod utilities;
-
-// Parser combinators.
-pub use empty_or::empty_or;
-pub use typed_list::typed_list;
-
-// Utility parser combinators.
-#[allow(unused_imports)]
-pub(crate) use utilities::{parens, prefix_expr, space_separated_list0, space_separated_list1, ws};
+pub mod utility;
